+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Changes:
 * Fix connection failure when sending too many environment variables (fix [#22](https://github.com/moul/sshportal/issues/22))
 * Fix panic when entering empty command (fix [#13](https://github.com/moul/sshportal/issues/13))
 * Add `config backup --ignore-events` option
+* Add `sshportal healthcheck [--addr=] [--wait] [--quiet]` cli command
+* Add [Docker Healthcheck](https://docs.docker.com/engine/reference/builder/#healthcheck) helper
 
 ## v1.6.0 (2017-12-12)